requirements.txt
# SynErgi - Multi-Agent RL Grid Optimization System
# Requirements file for Python dependencies

# Weave & W&B for LLM hosting and tracing
weave>=0.50.0
wandb>=0.16.0

# LangChain & LangGraph for multi-agent orchestration
langchain>=0.1.0
langgraph>=0.0.40
langchain-core>=0.1.0

# Environment variables
python-dotenv>=1.0.0

# LLM inference
openai>=1.0.0  # For OpenAI-compatible API calls to Qwen
langchain-openai>=0.0.5  # LangChain OpenAI integration

# Utilities
pydantic>=2.0.0  # For structured outputs
jsonschema
numpy

# Note: Qwen 2.5 will be accessed via Weave/W&B API
